+++ CHANGELOG.md
@@ -8,7 +8,9 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
 
 ##### Enhancements
 
-* None.  
+* Use the colored2 gem instead of colored.  
+  [Orta Therox](https://github.com/orta)
+  [xcodeproj#463](https://github.com/CocoaPods/Xcodeproj/pull/463)
 
 ##### Bug Fixes